+++ frontend/app/profile/account/page.tsx
@@ -4,7 +4,7 @@ import { Button, Divider, Input } from '@nextui-org/react';
 import RootLayout from '../../layout';
 import ProfileLayout from '../layout';
 import { Controller, useForm } from 'react-hook-form';
-import { PUT } from '../../axios/axios';
+import { PUT } from '../../libs/axios/axios';
 import useAuth from '../../hook/useAuth';
 import { notifyError, notifySuccess } from '../../components/Notifications';
 import DeleteAccountModal from '../../components/modal/deleteAccountModal';
