+++ frontend/app/questions/modal/addQuestionModal.tsx
@@ -16,7 +16,7 @@ import { Textarea } from '@nextui-org/react';
 import { Category, Complexity, Question } from '../../types/question';
 import { useForm } from 'react-hook-form';
 import { notifySuccess, notifyError } from '../../components/toast/notifications';
-import { createEntry } from '../../libs/axios/axios';
+import { POST } from '../../libs/axios/axios';
 
 
 export default function QuestionAddModal({fetchQuestions}) {
@@ -30,22 +30,21 @@ export default function QuestionAddModal({fetchQuestions}) {
     formState: { errors },
   } = useForm();
 
-  const onSubmit = handleSubmit((data: Question) => {
-      const modifiedData = {
-        ...data,
-        categories: (data.categories as string).split(',') as Category[],
-      };
-      createEntry('questions', modifiedData)
-      .then(res => {
-        if (res.message) {
-          fetchQuestions();
-          notifySuccess(res.message);
-          onOpenChange();
-          reset();
-        } else {
-          notifyError(res.error);
-        }
-      });
+  const onSubmit = handleSubmit(async (data: Question) => {
+    const modifiedData = {
+      ...data,
+      categories: (data.categories as string).split(',') as Category[],
+    };
+
+    try {
+      const response = await POST('questions', modifiedData);
+      fetchQuestions();
+      notifySuccess(response.data);
+      onOpenChange();
+      reset();
+    } catch (error) {
+      notifyError(error.message.data);
+    }
   });
 
   return (
