+++ frontend/app/questions/modal/deleteConfirmationModal.tsx
@@ -10,25 +10,24 @@ import {
 } from '@nextui-org/react';
 import { DeleteIcon } from '../../../public/DeleteIcon';
 import { notifyWarning, notifyError } from '../../components/toast/notifications';
-import { deleteEntry } from '../../libs/axios/axios';
+import { DELETE } from '../../libs/axios/axios';
 
 
 const DeleteConfirmationModal = ({ title, id, fetchQuestions }) => {
   const { isOpen, onOpen, onClose, onOpenChange } = useDisclosure();
 
-  const handleDelete = () => {
-    deleteEntry(`questions/${id}`)
-    .then(res => {
-      if (res.message) {
-        fetchQuestions();
-        notifyWarning(res.message);
-      } else {
-        notifyError(res.error);
-      }
-    })
-    .finally(() => onClose());
+  const handleDelete = async () => {
+    try {
+      const response = await DELETE(`questions/${id}`);
+      fetchQuestions();
+      notifyWarning(response.data);
+    } catch (error) {
+      notifyError(error.message.data);
+    } finally {
+      onClose();
+    }
   };
-  
+
   return (
     <>
       <Tooltip content="Delete question">
